requirements.txt
-e .[standard]

# Type annotation
asgiref~=3.5.2

# Explicit optionals
wsproto~=1.1.0

# Packaging
twine~=4.0.1
wheel~=0.37.1

# Testing
autoflake~=1.4
black~=22.6.0
flake8~=3.9.2
isort~=5.10.1
pytest~=7.1.2
pytest-mock~=3.8.2
pytest-asyncio~=0.19.0
mypy~=0.971
types-click~=7.1.8
types-pyyaml~=6.0.11
trustme~=0.9.0
cryptography~=37.0.4
coverage~=6.4.2
coverage-conditional-plugin~=0.5.0
httpx~=0.23.0
watchgod~=0.8.2

# Documentation
mkdocs~=1.3.1
mkdocs-material~=8.3.9

setuptools~=63.3.0
fastapi~=0.81
pyjwt~=2.4.0
starlette~=0.19.1
ujson~=5.4.0
